+++ src/modules/browse.ts
@@ -1,47 +1,17 @@
-import axios from 'axios'
-import jwt from 'jsonwebtoken'
 import opn from 'opn'
-import R from 'ramda'
-import { clusterIdDomainInfix, publicEndpoint } from '../api/env'
-import { Headers } from '../lib/constants/Headers'
+import QRCode from 'qrcode-terminal'
 import { SessionManager } from '../api/session/SessionManager'
-// Doesn't seem to work with 'import', seems to return undefined for some reason ¯\_(ツ)_/¯
-const QRCode = require('qrcode-terminal') // eslint-disable-line @typescript-eslint/no-var-requires
+import { storeUrl } from '../api/storeUrl'
 
-const isSupportRole = (role: string): boolean => role?.startsWith('vtex.support-authority')
-
-const isSupportSession = (): boolean => {
-  const { token } = SessionManager.getSingleton()
-  const decoded = jwt.decode(token)
-  if (!decoded || typeof decoded === 'string') {
-    return false
-  }
-
-  return R.any(role => isSupportRole(role), decoded.roles as string[])
+interface BrowseOptions {
+  qr: boolean
 }
 
-const prepareSupportBrowser = async (account: string, workspace: string): Promise<string> => {
-  const { token } = SessionManager.getSingleton()
-
-  const uri = `https://${workspace}--${account}.${publicEndpoint()}/_v/private/support-login/prepare`
-  const response = await axios.get(uri, {
-    headers: {
-      [Headers.VTEX_ORIGINAL_CREDENTIAL]: token,
-    },
-  })
-  return response.data.oneTimeToken
-}
-
-export default async (endpointInput, { q, qr }) => {
+export default async (path: string, { qr }: BrowseOptions) => {
   const { account, workspace } = SessionManager.getSingleton()
-  let endpoint = endpointInput ?? ''
-  if (isSupportSession()) {
-    const token = await prepareSupportBrowser(account, workspace)
-    endpoint = `_v/private/support-login/login?token=${token}&returnUrl=/${endpoint}`
-  }
-  const uri = `https://${workspace}--${account}${clusterIdDomainInfix()}.${publicEndpoint()}/${endpoint}`
+  const uri = storeUrl({ account, workspace, path })
 
-  if (q || qr) {
+  if (qr) {
     QRCode.generate(uri, { small: true })
     return
   }